services/chat/pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
sqlmodel = "^0.0.6"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pydantic = "^1.9.1"
PyJWT = {extras = ["crypto"], version = "^2.4.0"}
psycopg2 = "^2.9.3"
SQLAlchemy-Utils = "^0.38.2"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
